Update biz.aQute.junit to add AssertJ support

+++ biz.aQute.bndlib/src/aQute/bnd/osgi/EmbeddedResource.java
@@ -27,7 +27,7 @@ public class EmbeddedResource implements Resource {
 		this.lastModified = lastModified;
 	}
 
-	public EmbeddedResource(String pc, int lastModified) {
+	public EmbeddedResource(String pc, long lastModified) {
 		this(pc.getBytes(UTF_8), lastModified);
 	}
